+++ lib/kb_ObjectUtilities/kb_ObjectUtilitiesClient.py
@@ -133,6 +133,21 @@ class kb_ObjectUtilities(object):
         return self._client.call_method('kb_ObjectUtilities.KButil_update_genome_fields_from_files',
                                         [params], self._service_ver, context)
 
+    def KButil_update_genome_features_from_file(self, params, context=None):
+        """
+        :param params: instance of type
+           "KButil_update_genome_features_from_file_Params"
+           (KButil_update_genome_features_from_file() ** **  Method for
+           adding values to Genome object features, from file) -> structure:
+           parameter "feature_update_file" of type "file_path", parameter
+           "test_genome_ref_map" of mapping from String to String
+        :returns: instance of type
+           "KButil_update_genome_features_from_file_Output" -> structure:
+           parameter "updated_object_refs" of list of type "data_obj_ref"
+        """
+        return self._client.call_method('kb_ObjectUtilities.KButil_update_genome_features_from_file',
+                                        [params], self._service_ver, context)
+
     def status(self, context=None):
         return self._client.call_method('kb_ObjectUtilities.status',
                                         [], self._service_ver, context)