+++ lib/kb_ObjectUtilities/kb_ObjectUtilitiesServer.py
@@ -354,6 +354,10 @@ class Application(object):
                              name='kb_ObjectUtilities.KButil_update_genome_fields_from_files',
                              types=[dict])
         self.method_authentication['kb_ObjectUtilities.KButil_update_genome_fields_from_files'] = 'required'  # noqa
+        self.rpc_service.add(impl_kb_ObjectUtilities.KButil_update_genome_features_from_file,
+                             name='kb_ObjectUtilities.KButil_update_genome_features_from_file',
+                             types=[dict])
+        self.method_authentication['kb_ObjectUtilities.KButil_update_genome_features_from_file'] = 'required'  # noqa
         self.rpc_service.add(impl_kb_ObjectUtilities.status,
                              name='kb_ObjectUtilities.status',
                              types=[dict])