+++ lib/src/us/kbase/kbobjectutilities/KbObjectUtilitiesClient.java
@@ -232,6 +232,23 @@ public class KbObjectUtilitiesClient {
         return res.get(0);
     }
 
+    /**
+     * <p>Original spec-file function name: KButil_update_genome_features_from_file</p>
+     * <pre>
+     * </pre>
+     * @param   params   instance of type {@link us.kbase.kbobjectutilities.KButilUpdateGenomeFeaturesFromFileParams KButilUpdateGenomeFeaturesFromFileParams} (original type "KButil_update_genome_features_from_file_Params")
+     * @return   instance of type {@link us.kbase.kbobjectutilities.KButilUpdateGenomeFeaturesFromFileOutput KButilUpdateGenomeFeaturesFromFileOutput} (original type "KButil_update_genome_features_from_file_Output")
+     * @throws IOException if an IO exception occurs
+     * @throws JsonClientException if a JSON RPC exception occurs
+     */
+    public KButilUpdateGenomeFeaturesFromFileOutput kButilUpdateGenomeFeaturesFromFile(KButilUpdateGenomeFeaturesFromFileParams params, RpcContext... jsonRpcContext) throws IOException, JsonClientException {
+        List<Object> args = new ArrayList<Object>();
+        args.add(params);
+        TypeReference<List<KButilUpdateGenomeFeaturesFromFileOutput>> retType = new TypeReference<List<KButilUpdateGenomeFeaturesFromFileOutput>>() {};
+        List<KButilUpdateGenomeFeaturesFromFileOutput> res = caller.jsonrpcCall("kb_ObjectUtilities.KButil_update_genome_features_from_file", args, retType, true, true, jsonRpcContext, this.serviceVersion);
+        return res.get(0);
+    }
+
     public Map<String, Object> status(RpcContext... jsonRpcContext) throws IOException, JsonClientException {
         List<Object> args = new ArrayList<Object>();
         TypeReference<List<Map<String, Object>>> retType = new TypeReference<List<Map<String, Object>>>() {};