+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 - added method and app KButil_count_ws_objects()
 - added method and app KButil_update_genome_species_name()
 - added method (no app) KButil_update_genome_fields_from_files()
+- added method (no app) KButil_update_genome_features_from_files()
 
 ### Version 1.0.0
 - Initial release version